requirements.txt
pytest~=8.4.2
selene == 2.0.0rc9
selenium==4.31.0
allure-pytest
python-dotenv~=1.2.1
allure-python-commons~=2.15.0